rate_my_music/requirements.txt
asgiref==3.8.1
asttokens==2.4.1
colorama==0.4.6
decorator==5.1.1
Django==4.2
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-haystack==3.2.1
executing==2.0.1
ipython==8.23.0
jedi==0.19.1
matplotlib-inline==0.1.7
parso==0.8.4
pillow==10.3.0
prompt-toolkit==3.0.43
pure-eval==0.2.2
Pygments==2.17.2
python-dotenv==1.0.1
six==1.16.0
sqlparse==0.5.0
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.11.0
tzdata==2024.1
Unidecode==1.3.8
wcwidth==0.2.13
Whoosh==2.7.4
